+++ src/main/java/com/example/gimmegonghakauth/common/constant/AbeekTypeConst.java
@@ -1,4 +1,4 @@
-package com.example.gimmegonghakauth.constant;
+package com.example.gimmegonghakauth.common.constant;
 
 public enum AbeekTypeConst {
