+++ src/main/java/com/example/gimmegonghakauth/common/controller/MainController.java
@@ -1,4 +1,4 @@
-package com.example.gimmegonghakauth.controller;
+package com.example.gimmegonghakauth.common.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
