+++ src/main/java/com/example/gimmegonghakauth/common/infrastructure/MajorsDao.java
@@ -1,6 +1,6 @@
-package com.example.gimmegonghakauth.dao;
+package com.example.gimmegonghakauth.common.infrastructure;
 
-import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.common.domain.MajorsDomain;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
