+++ src/main/java/com/example/gimmegonghakauth/status/domain/AbeekDomain.java
@@ -1,7 +1,7 @@
 package com.example.gimmegonghakauth.status.domain;
 
-import com.example.gimmegonghakauth.constant.AbeekTypeConst;
-import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.common.constant.AbeekTypeConst;
+import com.example.gimmegonghakauth.common.domain.MajorsDomain;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
