+++ src/main/java/com/example/gimmegonghakauth/status/domain/GonghakCoursesDomain.java
@@ -1,8 +1,8 @@
 package com.example.gimmegonghakauth.status.domain;
 
-import com.example.gimmegonghakauth.constant.CourseCategoryConst;
-import com.example.gimmegonghakauth.domain.CoursesDomain;
-import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.common.constant.CourseCategoryConst;
+import com.example.gimmegonghakauth.common.domain.CoursesDomain;
+import com.example.gimmegonghakauth.common.domain.MajorsDomain;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
