+++ src/main/java/com/example/gimmegonghakauth/status/infrastructure/GonghakCoursesDao.java
@@ -1,8 +1,8 @@
 package com.example.gimmegonghakauth.status.infrastructure;
 
-import com.example.gimmegonghakauth.constant.CourseCategoryConst;
+import com.example.gimmegonghakauth.common.constant.CourseCategoryConst;
 import com.example.gimmegonghakauth.status.domain.GonghakCoursesDomain;
-import com.example.gimmegonghakauth.domain.MajorsDomain;
+import com.example.gimmegonghakauth.common.domain.MajorsDomain;
 import com.example.gimmegonghakauth.status.service.dto.GonghakCoursesByMajorDto;
 import com.example.gimmegonghakauth.status.service.dto.IncompletedCoursesDto;
 import java.util.List;
