+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakResultDto.java
@@ -1,6 +1,6 @@
 package com.example.gimmegonghakauth.status.service.dto;
 
-import com.example.gimmegonghakauth.constant.AbeekTypeConst;
+import com.example.gimmegonghakauth.common.constant.AbeekTypeConst;
 import java.util.Map;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
