+++ src/main/java/com/example/gimmegonghakauth/status/service/dto/GonghakStandardDto.java
@@ -1,7 +1,7 @@
 package com.example.gimmegonghakauth.status.service.dto;
 
 
-import com.example.gimmegonghakauth.constant.AbeekTypeConst;
+import com.example.gimmegonghakauth.common.constant.AbeekTypeConst;
 import java.util.Map;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
